app/requirements.txt
pandas==1.4.1
mysql-connector-python==8.0.28
requests==2.27.1
python-dotenv==0.20.0
streamlit>=1.0.0
matplotlib>=3.1.1
scikit-learn>=0.20.3
numpy==1.22.3
Flask==2.0.3
